agent-supervisor/requirements.txt
python-dotenv
langgraph
langgraph-supervisor
langchain-tavily
langchain[openai]
langgraph_supervisor